requirements.txt
aiogram==3.4.1
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
pytz==2023.3 
